+++ com.unity.mobile.notifications/Runtime/Unified/NotificationCenter.cs
@@ -85,7 +85,7 @@ namespace Unity.Notifications
         /// A custom non-empty string to identify notification channel. Required, Android only.
         /// All notifications will be sent to this channel.
         /// Channel is created automatically during initialization if <see cref="AndroidChannelName"/> and <see cref="AndroidChannelDescription"/> are both set.
-        /// If name and description are left null, channel with given identifier has to be created manually (for example using <see cref="AndroidNotificationCenter.RegisterNotificationChannel(AndroidNotificationChannel)"/>).
+        /// If name and description are left null, channel with given identifier has to be created manually (for example using <see cref="Unity.Notifications.Android.AndroidNotificationCenter.RegisterNotificationChannel(AndroidNotificationChannel)"/>).
         /// </summary>
         public string AndroidChannelId { get; set; }
 
@@ -106,6 +106,12 @@ namespace Unity.Notifications
         /// </summary>
         /// <seealso cref="AndroidChannelId"/>
         public string AndroidChannelDescription { get; set; }
+
+        /// <summary>
+        /// Small icon to be used for all notifications on Android. Optional, application icon will be used if not set.
+        /// This value will be assigned to <see cref="Unity.Notifications.Android.AndroidNotification.SmallIcon"/>.
+        /// </summary>
+        public string AndroidSmallIcon { get; set; }
     }
 
     /// <summary>
@@ -275,6 +281,7 @@ namespace Unity.Notifications
 #if UNITY_ANDROID
             var n = (AndroidNotification)notification;
             schedule.Schedule(ref n);
+            n.SmallIcon = s_Args.AndroidSmallIcon;
             if (notification.Identifier.HasValue)
             {
                 AndroidNotificationCenter.SendNotificationWithExplicitID(n, category, notification.Identifier.Value);